Switch to using string buffers in more places
By using string buffers (struct StringBuf), the likelyhood of buffer
overflow is reduced, and the number of places that bounds calculations
must be made is also reduced.  Furthermore, UTF-8 to UTF-16LE
conversion has been centralized, and the dependence on the printf()
family of functions has been eliminated.  This reduces the number of
compile errors when cross-compiling for Windows.