Fix intermittent hang on Windows
It is absolutely imperative that there be ONLY ONE instance of the named pipe
open but not connected when ConnectNamedPipe() is called.  Otherwise, Windows
will pick an arbitrary one.  This causes ConnectNamedPipe() to hang
intermittently, causing problems that are terribly difficult to debug!

The solution is to wait until ConnectNamedPipe() returns to make the next
CreateNamedPipe() call.  This bug took a very long time to fix.