Build system changes.  Finish async API.
This finishes the async API, but not all of it is implemented!  Also,
build system changes ensure that warning flags are actually passed to
the compiler, and that hidden visibility is used where available.
Furthermore, more features are tested with explicit feature tests,
instead of testing based on OS versions or compiler predefined macros.